pyproject.toml
[project]
name = "mistral-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mistralai>=1.5.1",
    "parse>=1.20.2",
    "python-dotenv>=1.0.1",
]
